[luci-interpreter] Support CWQ convolution op  (#4917)
Support convolution with channel-wise quantized weights and
layer-wise quantized activations (input and output of conv layer)

ONE-DCO-1.0-Signed-off-by: Alexander Efimov <a.efimov@samsung.com>

+++ compiler/luci-interpreter/include/luci_interpreter/core/Tensor.h
@@ -101,7 +101,7 @@ public:
     return _quantization.scale[0];
   }
 
-  float zero_point() const
+  int32_t zero_point() const
   {
     assert(_quantization.zero_point.size() == 1);
     return _quantization.zero_point[0];
